requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
sqlalchemy==2.0.43
pydantic==2.11.9
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.20
aiosqlite==0.21.0
asyncpg==0.29.0
python-dotenv==1.1.1
